requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
sqlalchemy==2.0.25
pandas==2.2.0
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
